Move plugin tests to test/plugins/index.js
